Use scripts dir only for WebUI
