+++ __init__.py
@@ -1,8 +1,8 @@
-from .esds_scripts import smea_utils
-from .esds_scripts.smea_sampling import sample_euler_dy, sample_euler_smea_dy
+from . import smea_sampling
+from .smea_sampling import sample_euler_dy, sample_euler_smea_dy
 
-if smea_utils.BACKEND == "ComfyUI":
-    if not smea_utils.INITIALIZED:
+if smea_sampling.BACKEND == "ComfyUI":
+    if not smea_sampling.INITIALIZED:
         from comfy.k_diffusion import sampling as k_diffusion_sampling
         from comfy.samplers import SAMPLER_NAMES
 
@@ -12,6 +12,6 @@ if smea_utils.BACKEND == "ComfyUI":
         SAMPLER_NAMES.append("euler_dy")
         SAMPLER_NAMES.append("euler_smea_dy")
 
-        smea_utils.INITIALIZED = True
+        smea_sampling.INITIALIZED = True
 
 NODE_CLASS_MAPPINGS = {}
